Start refactoring dynamics and integration
