+++ environment.yml
@@ -17,3 +17,4 @@ dependencies:
   - pip
   - pip:
       - pysme-astro
+      - jaxoplanet
